+++ proxy/chunkreq.go
@@ -65,6 +65,10 @@ func ChunkedProxyRequest(c *gin.Context, u string, cfg *config.Config, mode stri
 
 	req.Header.Set("Transfer-Encoding", "chunked") // 确保设置分块传输编码
 	setRequestHeaders(c, req)
+	// 删除Conection Upgrade头, 避免与HTTP/2冲突(检查是否存在Upgrade头)
+	if req.Header.Get("Connection") == "Upgrade" {
+		req.Header.Del("Connection")
+	}
 	AuthPassThrough(c, cfg, req)
 
 	resp, err := client.Do(req)